+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+Client 1.0.6:
+ - Change encoder to use extend_from_slice().
+
 ## [1.0.5]
 
 ### Added